Rewrite pigtail for integration tests
